+++ src/loger.rs
@@ -1,0 +1,23 @@
+use lazy_static::lazy_static;
+use slog::{o, Drain, Logger};
+
+lazy_static! {
+    pub static ref STDOUT: Logger = stdout_logger();
+    pub static ref STDERR: Logger = stderr_logger();
+}
+
+fn stdout_logger() -> Logger {
+    let decorator = slog_term::TermDecorator::new().build();
+    let drain = slog_term::CompactFormat::new(decorator).build().fuse();
+    let drain = slog_async::Async::new(drain).build().fuse();
+
+    Logger::root(drain, o!())
+}
+
+fn stderr_logger() -> Logger {
+    let decorator = slog_term::TermDecorator::new().build();
+    let drain = slog_term::CompactFormat::new(decorator).build();
+    let drain = std::sync::Mutex::new(drain).fuse();
+
+    Logger::root(drain, o!())
+}