+++ src/main.rs
@@ -1,53 +1,35 @@
-use crate::app_config::AppConfig;
 use crate::config::{Config, ConfigBuilder};
 use crate::handler::Handler;
+use crate::loger::{STDERR, STDOUT};
 use anyhow::Result;
 use clap::Parser;
 use failure::Error;
-use lazy_static::lazy_static;
 use option::Args;
 use slog::{crit, debug, info};
-use slog::{o, Drain, Logger};
 use std::fmt::Display;
 use std::fs::File;
 use std::io::prelude::*;
-use std::net::SocketAddr;
 use std::process::exit;
 use std::time::Duration;
 use tokio;
 use tokio::net::{TcpListener, UdpSocket};
 use trust_dns_server::ServerFuture;
 
-mod app_config;
 mod config;
 mod domain;
 mod filter;
 mod handler;
+mod handler_config;
 mod ip;
+mod loger;
 mod option;
 mod resolver;
 
-struct MainConfig {
-    pub bind: SocketAddr,
-    pub app_config: AppConfig,
-}
-
-lazy_static! {
-    static ref STDOUT: Logger = stdout_logger();
-    static ref STDERR: Logger = stderr_logger();
-    static ref CONFIG: MainConfig = {
-        let config = config().unwrap_or_log();
-        MainConfig {
-            bind: config.bind,
-            app_config: AppConfig::new(config),
-        }
-    };
-}
-
 #[tokio::main]
 async fn main() -> Result<()> {
-    let bind_socket = CONFIG.bind;
-    let mut server = ServerFuture::new(Handler::default());
+    let config = config().unwrap_or_log();
+    let bind_socket = config.bind;
+    let mut server = ServerFuture::new(Handler::new(config.into()));
 
     let bind = UdpSocket::bind(bind_socket);
     info!(STDOUT, "Listening on UDP: {}", bind_socket);
@@ -71,22 +53,6 @@ fn config() -> Result<Config, Error> {
 
     let builder: ConfigBuilder = toml::from_str(&content)?;
     builder.build()
-}
-
-fn stdout_logger() -> Logger {
-    let decorator = slog_term::TermDecorator::new().build();
-    let drain = slog_term::CompactFormat::new(decorator).build().fuse();
-    let drain = slog_async::Async::new(drain).build().fuse();
-
-    Logger::root(drain, o!())
-}
-
-fn stderr_logger() -> Logger {
-    let decorator = slog_term::TermDecorator::new().build();
-    let drain = slog_term::CompactFormat::new(decorator).build();
-    let drain = std::sync::Mutex::new(drain).fuse();
-
-    Logger::root(drain, o!())
 }
 
 trait ShouldSuccess {
@@ -119,22 +85,5 @@ where
             debug!(STDERR, "{:?}", e.backtrace());
             exit(1);
         })
-    }
-}
-
-trait Transpose {
-    type Output;
-    fn transpose(self) -> Self::Output;
-}
-
-impl<T, E> Transpose for Option<Result<T, E>> {
-    type Output = Result<Option<T>, E>;
-
-    fn transpose(self) -> Self::Output {
-        match self {
-            Some(Ok(x)) => Ok(Some(x)),
-            Some(Err(e)) => Err(e),
-            None => Ok(None),
-        }
     }
 }